# iadmin-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Server pets",
        "title": "Swagger Example API",
        "contact": {},
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/sysRole/addSysRoleInfo": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SysRole"
                ],
                "summary": "新增角色",
                "parameters": [
                    {
                        "description": "角色模型",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SysRole"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"新增成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/sysRole/deleteBatchSysRole": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SysRole"
                ],
                "summary": "批量删除角色",
                "parameters": [
                    {
                        "description": "id列表",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.IdsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"批量删除成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/sysRole/excelOut": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SysRole"
                ],
                "summary": "批量导出角色",
                "parameters": [
                    {
                        "description": "批量导出请求模型",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SysRoleExcelOut"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"批量导出成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/sysRole/getSysRoleList": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SysRole"
                ],
                "summary": "分页获取角色列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/sysRole/setRoleDefaultRouter": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SysRole"
                ],
                "summary": "配置角色菜单首页",
                "parameters": [
                    {
                        "description": "role_id,default_router",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SysRoleDefaultRouter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"配置角色接口成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/sysRole/updateSysRoleInfo": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SysRole"
                ],
                "summary": "更新角色信息",
                "parameters": [
                    {
                        "description": "角色模型",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SysRole"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"更新成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/sysRole/updateSysRoleMenuConfig": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SysRole"
                ],
                "summary": "配置角色菜单",
                "parameters": [
                    {
                        "description": "role_id,id列表",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SysRoleMenuConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"配置角色菜单成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.SysMenu": {
            "type": "object",
            "properties": {
                "Component": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "full_page": {
                    "type": "boolean"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "keep_alive": {
                    "type": "boolean"
                },
                "menuLevel": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "parentId": {
                    "description": "父菜单ID",
                    "type": "integer"
                },
                "path": {
                    "description": "路由path",
                    "type": "string"
                },
                "show_in_menu": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
...
        "model.SysRole": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "default_router": {
                    "description": "首页路径(默认/index)",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SysMenu"
                    }
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
...
        "request.ExcelOutRequest": {
            "type": "object",
            "properties": {
                "hasAllData": {
                    "type": "boolean"
                },
                "hasTableHead": {
                    "type": "boolean"
                }
            }
        },
        "request.IdsReq": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "request.Pagination": {
            "type": "object",
            "properties": {
                "current": {
                    "description": "页码",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页大小",
                    "type": "integer"
                }
            }
        },
...
        "request.SysRoleDefaultRouter": {
            "type": "object",
            "properties": {
                "default_router": {
                    "type": "string"
                },
                "role_id": {
                    "type": "integer"
                }
            }
        },
        "request.SysRoleExcelOut": {
            "type": "object",
            "properties": {
                "excelOutConfig": {
                    "$ref": "#/definitions/request.ExcelOutRequest"
                },
                "pagination": {
                    "$ref": "#/definitions/request.Pagination"
                }
            }
        },
        "request.SysRoleMenuConfig": {
            "type": "object",
            "properties": {
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SysMenu"
                    }
                },
                "role_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}